Trying to make sure I know what direction I'm going in, I spotted a whole bundle of errors - some I can fix here, some I can't. I've documented what I can't in the file "Outstanding Tasks.php" - created as a PHP file so the TODO items it'll show up in the problems section of Eclipse.
Fixes so far:
* Corrected some typos in the ChartBroker. Added a new (dummy) function
to support one of the two new RSS feeds for the Chart. In time this will
be migrated to actually return useful information.
* Added OGA and M4A support to the HTML class for RSS feeds of the
internal shows. Added data to support the new RSS feeds for the Chart.
Some spacing fixes as identified by PHP_CodeSniffer
* Added a docblock to the LicenseSelector as spotted by PHP_CodeSniffer
* Noticed some PHP Errors in the errorlog around the NewUserObject. Also
spotted that using cURL to act against a non-authenticated API call
returns a new user each time. Tried to resolve it. Failed. More to come!
* Added docblock to RemoteSources, and corrected some object inheritance
here too.
* Added sensible caching to ShowBroker - by ShowID as well as ShowDate
(for internal shows).
* Corrected some spacing issues in TrackBroker as identified by
PHP_CodeSniffer.
* Noticed some errors in TrackObject and VoteBroker where there are no
shows created for a track.
* Added a missing function back into the UserObject relating to the
e-mail address supplied by the OpenID provider.
* Corrected some spacing in the Chart.html.tpl file. Also, tried to
remove some errors around page numbering (and the lack of passed page
number data) here too, without success.
* Added an A Name record to the show_track_data.tpl.
* Created Outstanding Tasks.php to document missing code.
* Added new RSS feeds for the Day's chart changes as a per-change item
and as a per-day item. Neither work the way I want them to right now.
* Added the templates for the .mp3, oga and .m4a encoded RSS feeds for
the shows.

Lots more changes to come over the next few days.

+++ CLASSES/class_RemoteSources.php
@@ -60,6 +60,11 @@ class RemoteSources extends GenericObject
     protected $fileName = "";
     protected $intUserID = 0;
 
+    /**
+     * Remove this RemoteSource from the Processing table.
+     *
+     * @return boolean Action completed.
+     */
     public function cancel()
     {
         $db = Database::getConnection();
@@ -67,6 +72,7 @@ class RemoteSources extends GenericObject
             $sql = "DELETE FROM processing WHERE intProcessingID = ?";
             $query = $db->prepare($sql);
             $query->execute(array($this->intProcessingID));
+            return true;
         } catch(Exception $e) {
             error_log("SQL Died: " . $e->getMessage());
             return false;
@@ -568,9 +574,9 @@ class RemoteSources extends GenericObject
                 $this->strArtistUrl
             );
             $this->set_intArtistID($artist->get_intArtistID());
-            $this->write();
+            parent::write();
         }
-        $this->intTrackID = new NewTrackObject(
+        $track = new NewTrackObject(
             $this->intArtistID,
             $this->strTrackName,
             $this->strTrackNameSounds,
@@ -579,11 +585,13 @@ class RemoteSources extends GenericObject
             $this->isNSFW,
             $this->fileName
         );
-        if ($this->fileUrl != '') {
+        if ($this->fileUrl != '' and $track != false) {
             unlink($this->fileName);
         }
-        $this->cancel();
-        return $this->intTrackID;
+        if ($track != false) {
+            $this->cancel();
+            return $track->get_intTrackID();
+        }
     }
 
     /**
